Gradle: Move tasks to its daemon's package
