+++ buildSrc/src/main/kotlin/bisq/gradle/bitcoind/BitcoindProcessTasks.kt
@@ -1,7 +1,7 @@
 package bisq.gradle.bitcoind
 
-import bisq.gradle.tasks.bitcoind.BitcoindStopTask
-import bisq.gradle.tasks.bitcoind.StartBitcoinQtTask
+import bisq.gradle.bitcoind.tasks.BitcoindStopTask
+import bisq.gradle.bitcoind.tasks.StartBitcoinQtTask
 import org.gradle.api.Project
 import org.gradle.api.tasks.Delete
 import org.gradle.api.tasks.TaskContainer
