+++ buildSrc/src/main/kotlin/bisq/gradle/bitcoind/BitcoindRegtestTasks.kt
@@ -1,7 +1,7 @@
 package bisq.gradle.bitcoind
 
-import bisq.gradle.tasks.ApplicationRunTaskFactory
-import bisq.gradle.tasks.bitcoind.StartBitcoinQtTask
+import bisq.gradle.ApplicationRunTaskFactory
+import bisq.gradle.bitcoind.tasks.StartBitcoinQtTask
 import org.gradle.api.Project
 import org.gradle.api.tasks.TaskProvider
 import java.io.File
