+++ buildSrc/src/main/kotlin/bisq/gradle/bitcoind/tasks/BitcoindStopTask.kt
@@ -1,19 +1,19 @@
-package bisq.gradle.tasks.elementsd
+package bisq.gradle.bitcoind.tasks
 
 import bisq.gradle.Network
-import bisq.gradle.elementsd.ElementsdRpcClient
+import bisq.gradle.bitcoind.BitcoindRpcClient
 import org.gradle.api.DefaultTask
 import org.gradle.api.provider.Property
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.StopExecutionException
 import org.gradle.api.tasks.TaskAction
 
-abstract class ElementsdStopTask : DefaultTask() {
+abstract class BitcoindStopTask : DefaultTask() {
     @get:Input
     abstract val port: Property<Int>
 
     init {
-        port.convention(StartElementsQtTask.DEFAULT_ELEMENTSD_RPC_PORT)
+        port.convention(StartBitcoinQtTask.DEFAULT_BITCOIND_RPC_PORT)
     }
 
     @TaskAction
@@ -22,7 +22,7 @@ abstract class ElementsdStopTask : DefaultTask() {
             throw StopExecutionException("bitcoind is not running.")
         }
 
-        ElementsdRpcClient.daemonRpcCall(
+        BitcoindRpcClient.daemonRpcCall(
             listOf("stop")
         )
     }
